+++ metatensor-core/src/c_api/io/block.rs
@@ -9,62 +9,34 @@ use crate::data::mts_array_t;
 use super::{ExternalBuffer, mts_realloc_buffer_t};
 
 use super::super::status::{mts_status_t, catch_unwind};
-use super::super::tensor::mts_tensormap_t;
+use super::super::blocks::mts_block_t;
 use super::mts_create_array_callback_t;
 
-/// Load a tensor map from the file at the given path.
+
+/// Load a tensor block from the file at the given path.
 ///
 /// Arrays for the values and gradient data will be created with the given
 /// `create_array` callback, and filled by this function with the corresponding
 /// data.
 ///
 /// The memory allocated by this function should be released using
-/// `mts_tensormap_free`.
+/// `mts_block_free`.
 ///
-/// `TensorMap` are serialized using numpy's `.npz` format, i.e. a ZIP file
-/// without compression (storage method is STORED), where each file is stored as
-/// a `.npy` array. Both the ZIP and NPY format are well documented:
-///
-/// - ZIP: <https://pkware.cachefly.net/webdocs/casestudies/APPNOTE.TXT>
-/// - NPY: <https://numpy.org/doc/stable/reference/generated/numpy.lib.format.html>
-///
-/// We add other restriction on top of these formats when saving/loading data.
-/// First, `Labels` instances are saved as structured array, see the `labels`
-/// module for more information. Only 32-bit integers are supported for Labels,
-/// and only 64-bit floats are supported for data (values and gradients).
-///
-/// Second, the path of the files in the archive also carry meaning. The keys of
-/// the `TensorMap` are stored in `/keys.npy`, and then different blocks are
-/// stored as
-///
-/// ```bash
-/// /  blocks / <block_id>  / values / samples.npy
-///                         / values / components  / 0.npy
-///                                                / <...>.npy
-///                                                / <n_components>.npy
-///                         / values / properties.npy
-///                         / values / data.npy
-///
-///                         # optional sections for gradients, one by parameter
-///                         /   gradients / <parameter> / samples.npy
-///                                                     /   components  / 0.npy
-///                                                                     / <...>.npy
-///                                                                     / <n_components>.npy
-///                                                     /   data.npy
-/// ```
+/// See `mts_tensormap_load` for more information about the format used to
+/// serialize the data.
 ///
 /// @param path path to the file as a NULL-terminated UTF-8 string
 /// @param create_array callback function that will be used to create data
 ///                     arrays inside each block
 ///
-/// @returns A pointer to the newly allocated tensor map, or a `NULL` pointer in
+/// @returns A pointer to the newly allocated block, or a `NULL` pointer in
 ///          case of error. In case of error, you can use `mts_last_error()`
 ///          to get the error message.
 #[no_mangle]
-pub unsafe extern fn mts_tensormap_load(
+pub unsafe extern fn mts_block_load(
     path: *const c_char,
     create_array: mts_create_array_callback_t,
-) -> *mut mts_tensormap_t {
+) -> *mut mts_block_t {
     let mut result = std::ptr::null_mut();
     let unwind_wrapper = std::panic::AssertUnwindSafe(&mut result);
     let status = catch_unwind(move || {
@@ -74,20 +46,20 @@ pub unsafe extern fn mts_tensormap_load(
 
         let path = CStr::from_ptr(path).to_str().expect("use UTF-8 for path");
         let file = BufReader::new(File::open(path)?);
-        let tensor = crate::io::load(file, create_array)
+        let block = crate::io::load_block(file, create_array)
             .map_err(|err| match err {
                 Error::Serialization(message) => {
                     if crate::io::looks_like_labels_data(crate::io::PathOrBuffer::Path(path)) {
                         Error::Serialization(format!(
-                            "unable to load a TensorMap from '{}', use `load_labels` to load Labels: {}", path, message
+                            "unable to load a TensorBlock from '{}', use `load_labels` to load Labels: {}", path, message
                         ))
-                    } else if crate::io::looks_like_block_data(crate::io::PathOrBuffer::Path(path)) {
+                    } else if crate::io::looks_like_tensormap_data(crate::io::PathOrBuffer::Path(path)) {
                         Error::Serialization(format!(
-                            "unable to load a TensorMap from '{}', use `load_block` to load TensorBlock: {}", path, message
+                            "unable to load a TensorBlock from '{}', use `load` to load TensorMap: {}", path, message
                         ))
                     } else {
                         Error::Serialization(format!(
-                            "unable to load a TensorMap from '{}': {}", path, message
+                            "unable to load a TensorBlock from '{}': {}", path, message
                         ))
                     }
                 }
@@ -97,7 +69,7 @@ pub unsafe extern fn mts_tensormap_load(
         // force the closure to capture the full unwind_wrapper, not just
         // unwind_wrapper.0
         let _ = &unwind_wrapper;
-        *(unwind_wrapper.0) = mts_tensormap_t::into_boxed_raw(tensor);
+        *(unwind_wrapper.0) = mts_block_t::into_boxed_raw(block);
         Ok(())
     });
 
@@ -108,29 +80,29 @@ pub unsafe extern fn mts_tensormap_load(
     return result;
 }
 
-/// Load a tensor map from the given in-memory buffer.
+/// Load a tensor block from the given in-memory buffer.
 ///
 /// Arrays for the values and gradient data will be created with the given
 /// `create_array` callback, and filled by this function with the corresponding
 /// data.
 ///
 /// The memory allocated by this function should be released using
-/// `mts_tensormap_free`.
+/// `mts_block_free`.
 ///
-/// @param buffer buffer containing a previously serialized `mts_tensormap_t`
+/// @param buffer buffer containing a previously serialized `mts_block_t`
 /// @param buffer_count number of elements in the buffer
 /// @param create_array callback function that will be used to create data
 ///                     arrays inside each block
 ///
-/// @returns A pointer to the newly allocated tensor map, or a `NULL` pointer in
-///          case of error. In case of error, you can use `mts_last_error()`
+/// @returns A pointer to the newly allocated tensor block, or a `NULL` pointer
+///          in case of error. In case of error, you can use `mts_last_error()`
 ///          to get the error message.
 #[no_mangle]
-pub unsafe extern fn mts_tensormap_load_buffer(
+pub unsafe extern fn mts_block_load_buffer(
     buffer: *const u8,
     buffer_count: usize,
     create_array: mts_create_array_callback_t,
-) -> *mut mts_tensormap_t {
+) -> *mut mts_block_t {
     let mut result = std::ptr::null_mut();
     let unwind_wrapper = std::panic::AssertUnwindSafe(&mut result);
     let status = catch_unwind(move || {
@@ -142,22 +114,22 @@ pub unsafe extern fn mts_tensormap_load_buffer(
         let slice = std::slice::from_raw_parts(buffer.cast::<u8>(), buffer_count);
         let cursor = std::io::Cursor::new(slice);
 
-        let tensor = crate::io::load(cursor, create_array)
+        let block = crate::io::load_block(cursor, create_array)
             .map_err(|err| match err {
                 Error::Serialization(message) => {
                     let slice = std::slice::from_raw_parts(buffer.cast::<u8>(), buffer_count);
                     let mut cursor = std::io::Cursor::new(slice);
                     if crate::io::looks_like_labels_data(crate::io::PathOrBuffer::Buffer(&mut cursor)) {
                         Error::Serialization(format!(
-                            "unable to load a TensorMap from buffer, use `load_labels_buffer` to load Labels: {}", message
+                            "unable to load a TensorBlock from buffer, use `load_labels` to load Labels: {}", message
                         ))
-                    } else if crate::io::looks_like_block_data(crate::io::PathOrBuffer::Buffer(&mut cursor)) {
+                    } else if crate::io::looks_like_tensormap_data(crate::io::PathOrBuffer::Buffer(&mut cursor)) {
                         Error::Serialization(format!(
-                            "unable to load a TensorMap from buffer, use `load_block_buffer` to load TensorBlock: {}", message
+                            "unable to load a TensorBlock from buffer, use `load` to load TensorMap: {}", message
                         ))
                     } else {
                         Error::Serialization(format!(
-                            "unable to load a TensorMap from buffer: {}", message
+                            "unable to load a TensorBlock from buffer: {}", message
                         ))
                     }
                 }
@@ -167,7 +139,7 @@ pub unsafe extern fn mts_tensormap_load_buffer(
         // force the closure to capture the full unwind_wrapper, not just
         // unwind_wrapper.0
         let _ = &unwind_wrapper;
-        *(unwind_wrapper.0) = mts_tensormap_t::into_boxed_raw(tensor);
+        *(unwind_wrapper.0) = mts_block_t::into_boxed_raw(block);
         Ok(())
     });
 
@@ -194,41 +166,41 @@ fn wrap_create_array(create_array: &mts_create_array_callback_t) -> impl Fn(Vec<
         } else {
             return Err(Error::External {
                 status: status,
-                context: "failed to create a new array in mts_tensormap_load".into()
+                context: "failed to create a new array in mts_block_load".into()
             });
         }
     }
 }
 
 
-/// Save a tensor map to the file at the given path.
+/// Save a tensor block to the file at the given path.
 ///
 /// If the file already exists, it is overwritten.
 ///
 /// @param path path to the file as a NULL-terminated UTF-8 string
-/// @param tensor tensor map to save to the file
+/// @param block tensor block to save to the file
 ///
 /// @returns The status code of this operation. If the status is not
 ///          `MTS_SUCCESS`, you can use `mts_last_error()` to get the full
 ///          error message.
 #[no_mangle]
-pub unsafe extern fn mts_tensormap_save(
+pub unsafe extern fn mts_block_save(
     path: *const c_char,
-    tensor: *const mts_tensormap_t,
+    block: *const mts_block_t,
 ) -> mts_status_t {
     catch_unwind(|| {
-        check_pointers_non_null!(path, tensor);
+        check_pointers_non_null!(path, block);
 
         let path = CStr::from_ptr(path).to_str().expect("use UTF-8 for path");
         let file = BufWriter::new(File::create(path)?);
-        crate::io::save(file, &*tensor)?;
+        crate::io::save_block(file, &*block)?;
 
         Ok(())
     })
 }
 
 
-/// Save a tensor map to an in-memory buffer.
+/// Save a tensor block to an in-memory buffer.
 ///
 /// On input, `*buffer` should contain the address of a starting buffer (which
 /// can be NULL) and `*buffer_count` should contain the size of the allocation.
@@ -240,33 +212,33 @@ pub unsafe extern fn mts_tensormap_save(
 /// Users of this function are responsible for freeing the `*buffer` when they
 /// are done with it, using the function matching the `realloc` callback.
 ///
-/// @param buffer pointer to the buffer the tensor will be stored to, which can
+/// @param buffer pointer to the buffer the block will be stored to, which can
 ///        change due to reallocations.
 /// @param buffer_count pointer to the buffer size on input, number of written
 ///        bytes on output
 /// @param realloc_user_data custom data for the `realloc` callback. This will
 ///        be passed as the first argument to `realloc` as-is.
 /// @param realloc function that allows to grow the buffer allocation
-/// @param tensor tensor map that will saved to the buffer
+/// @param block tensor block that will saved to the buffer
 ///
 /// @returns The status code of this operation. If the status is not
 ///          `MTS_SUCCESS`, you can use `mts_last_error()` to get the full error
 ///          message.
 #[no_mangle]
 #[allow(clippy::cast_possible_truncation)]
-pub unsafe extern fn mts_tensormap_save_buffer(
+pub unsafe extern fn mts_block_save_buffer(
     buffer: *mut *mut u8,
     buffer_count: *mut usize,
     realloc_user_data: *mut c_void,
     realloc: mts_realloc_buffer_t,
-    tensor: *const mts_tensormap_t,
+    block: *const mts_block_t,
 ) -> mts_status_t {
     catch_unwind(|| {
-        check_pointers_non_null!(tensor, buffer_count, buffer);
+        check_pointers_non_null!(block, buffer_count, buffer);
 
         if realloc.is_none() {
             return Err(Error::InvalidParameter(
-                "realloc callback can not be NULL in mts_tensormap_save_buffer".into()
+                "realloc callback can not be NULL in mts_block_save_buffer".into()
             ));
         }
 
@@ -282,7 +254,7 @@ pub unsafe extern fn mts_tensormap_save_buffer(
             current: 0,
         };
 
-        crate::io::save(&mut external_buffer, &*tensor)?;
+        crate::io::save_block(&mut external_buffer, &*block)?;
 
         *buffer_count = external_buffer.current as usize;