+++ python/metatensor-torch/tests/serialization.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 import metatensor.torch
-from metatensor.torch import Labels, TensorMap
+from metatensor.torch import Labels, TensorBlock, TensorMap
 
 from . import _tests_utils
 
@@ -21,6 +21,19 @@ def tensor_path():
         "metatensor-core",
         "tests",
         "data.npz",
+    )
+
+
+@pytest.fixture
+def block_path():
+    return os.path.join(
+        os.path.dirname(__file__),
+        "..",
+        "..",
+        "..",
+        "metatensor-core",
+        "tests",
+        "block.npz",
     )
 
 
@@ -47,10 +60,7 @@ def check_labels(labels):
     assert len(labels) == 27
 
 
-def check_tensor(tensor):
-    check_labels(tensor.keys)
-
-    block = tensor.block(dict(o3_lambda=2, center_type=6, neighbor_type=1))
+def check_block(block):
     assert block.samples.names == ["system", "atom"]
     assert block.values.shape == (9, 5, 3)
 
@@ -59,25 +69,31 @@ def check_tensor(tensor):
     assert gradient.values.shape == (59, 3, 5, 3)
 
 
+def check_tensor(tensor):
+    check_labels(tensor.keys)
+
+    block = tensor.block(dict(o3_lambda=2, center_type=6, neighbor_type=1))
+    check_block(block)
+
+
 def test_load(tensor_path):
     loaded = metatensor.torch.load(tensor_path)
     check_tensor(loaded)
 
-    loaded = metatensor.torch.TensorMap.load(tensor_path)
+    loaded = TensorMap.load(tensor_path)
     check_tensor(loaded)
 
-
-def test_load_buffer(tensor_path):
+    # load from buffer
     buffer = torch.tensor(np.fromfile(tensor_path, dtype="uint8"))
 
     loaded = metatensor.torch.load_buffer(buffer)
     check_tensor(loaded)
 
-    loaded = metatensor.torch.TensorMap.load_buffer(buffer)
+    loaded = TensorMap.load_buffer(buffer)
     check_tensor(loaded)
 
 
-def test_save(tmpdir):
+def test_save(tmpdir, tensor_path):
     """Check that we can save and load a tensor to a file"""
     tmpfile = "serialize-test.npz"
 
@@ -94,9 +110,7 @@ def test_save(tmpdir):
 
         assert len(data.keys) == 4
 
-
-def test_save_buffer(tensor_path):
-    """Check that we can save and load a tensor to an in-memory buffer"""
+    # save using buffer
     buffer = torch.tensor(np.fromfile(tensor_path, dtype="uint8"))
     tensor = metatensor.torch.load_buffer(buffer)
 
@@ -131,6 +145,60 @@ def test_save_load_zero_length_block(tmpdir):
         metatensor.torch.load(file)
 
 
+def test_load_block(block_path):
+    loaded = metatensor.torch.load_block(block_path)
+    check_block(loaded)
+
+    loaded = TensorBlock.load(block_path)
+    check_block(loaded)
+
+    # load from buffer
+    buffer = torch.tensor(np.fromfile(block_path, dtype="uint8"))
+
+    loaded = metatensor.torch.load_block_buffer(buffer)
+    check_block(loaded)
+
+    loaded = TensorBlock.load_buffer(buffer)
+    check_block(loaded)
+
+
+def test_save_block(tmpdir, block_path):
+    """Check that we can save and load a tensor to a file"""
+    tmpfile = "serialize-test.npz"
+
+    block = _tests_utils.tensor(dtype=torch.float64).block(2)
+
+    with tmpdir.as_cwd():
+        metatensor.torch.save(tmpfile, block)
+        data = metatensor.torch.load_block(tmpfile)
+        assert len(data) == 4
+
+        block.save(tmpfile)
+        data = metatensor.torch.load_block(tmpfile)
+        assert len(data) == 4
+
+    # save with buffer
+    buffer = torch.tensor(np.fromfile(block_path, dtype="uint8"))
+    tensor = metatensor.torch.load_block_buffer(buffer)
+
+    saved = metatensor.torch.save_buffer(tensor)
+    assert torch.all(buffer == saved)
+
+    saved = tensor.save_buffer()
+    assert torch.all(buffer == saved)
+
+
+def test_pickle_block(tmpdir, block_path):
+    block = metatensor.torch.load_block(block_path)
+    tmpfile = "serialize-test.npz"
+
+    with tmpdir.as_cwd():
+        torch.save(block, tmpfile)
+        loaded = torch.load(tmpfile)
+
+    check_block(loaded)
+
+
 def test_load_labels(labels_path):
     loaded = metatensor.torch.load_labels(labels_path)
     check_labels(loaded)
@@ -138,20 +206,19 @@ def test_load_labels(labels_path):
     loaded = metatensor.torch.Labels.load(labels_path)
     check_labels(loaded)
 
-
-def test_load_labels_buffer(labels_path):
+    # load from buffer
     buffer = torch.tensor(np.fromfile(labels_path, dtype="uint8"))
 
     loaded = metatensor.torch.load_labels_buffer(buffer)
     check_labels(loaded)
 
-    loaded = metatensor.torch.Labels.load_buffer(buffer)
+    loaded = Labels.load_buffer(buffer)
     check_labels(loaded)
 
 
-def test_save_labels(tmpdir):
+def test_save_labels(tmpdir, labels_path):
     """Check that we can save and load a tensor to a file"""
-    tmpfile = "serialize-test.npz"
+    tmpfile = "serialize-test.npy"
 
     labels = _tests_utils.tensor(dtype=torch.float64).keys
 
@@ -164,8 +231,7 @@ def test_save_labels(tmpdir):
         data = metatensor.torch.load_labels(tmpfile)
         assert len(data) == 4
 
-
-def test_save_labels_buffer(labels_path):
+    # save with buffer
     buffer = torch.tensor(np.fromfile(labels_path, dtype="uint8"))
     tensor = metatensor.torch.load_labels_buffer(buffer)
 
@@ -177,11 +243,11 @@ def test_save_labels_buffer(labels_path):
 
 
 def test_pickle_labels(tmpdir, labels_path):
-    tensor = metatensor.torch.load_labels(labels_path)
-    tmpfile = "serialize-test.npz"
+    labels = metatensor.torch.load_labels(labels_path)
+    tmpfile = "serialize-test.npy"
 
     with tmpdir.as_cwd():
-        torch.save(tensor, tmpfile)
+        torch.save(labels, tmpfile)
         loaded = torch.load(tmpfile)
 
     check_labels(loaded)
@@ -194,16 +260,22 @@ class Serialization:
     def load_buffer(self, buffer: torch.Tensor) -> TensorMap:
         return metatensor.torch.load_buffer(buffer=buffer)
 
+    def load_block(self, path: str) -> TensorBlock:
+        return metatensor.torch.load_block(path=path)
+
+    def load_block_buffer(self, buffer: torch.Tensor) -> TensorBlock:
+        return metatensor.torch.load_block_buffer(buffer=buffer)
+
     def load_labels(self, path: str) -> Labels:
         return metatensor.torch.load_labels(path=path)
 
     def load_labels_buffer(self, buffer: torch.Tensor) -> Labels:
         return metatensor.torch.load_labels_buffer(buffer=buffer)
 
-    def save(self, path: str, data: Union[Labels, TensorMap]):
+    def save(self, path: str, data: Union[Labels, TensorBlock, TensorMap]):
         return metatensor.torch.save(path=path, data=data)
 
-    def save_buffer(self, data: Union[Labels, TensorMap]) -> torch.Tensor:
+    def save_buffer(self, data: Union[Labels, TensorBlock, TensorMap]) -> torch.Tensor:
         return metatensor.torch.save_buffer(data=data)