+++ rust/metatensor/tests/serialization.rs
@@ -3,18 +3,20 @@ mod tensor {
 
     use metatensor::TensorMap;
 
+    const DATA_PATH: &str = "../../metatensor-core/tests/data.npz";
+
     #[test]
     fn load_file() {
-        let tensor = metatensor::io::load("../../metatensor-core/tests/data.npz").unwrap();
+        let tensor = metatensor::io::load(DATA_PATH).unwrap();
         check_tensor(&tensor);
 
-        let tensor = TensorMap::load("../../metatensor-core/tests/data.npz").unwrap();
+        let tensor = TensorMap::load(DATA_PATH).unwrap();
         check_tensor(&tensor);
     }
 
     #[test]
     fn load_buffer() {
-        let mut file = std::fs::File::open("../../metatensor-core/tests/data.npz").unwrap();
+        let mut file = std::fs::File::open(DATA_PATH).unwrap();
         let mut buffer = Vec::new();
         file.read_to_end(&mut buffer).unwrap();
 
@@ -27,7 +29,7 @@ mod tensor {
 
     #[test]
     fn save_buffer() {
-        let mut file = std::fs::File::open("../../metatensor-core/tests/data.npz").unwrap();
+        let mut file = std::fs::File::open(DATA_PATH).unwrap();
         let mut buffer = Vec::new();
         file.read_to_end(&mut buffer).unwrap();
 
@@ -65,23 +67,88 @@ mod tensor {
     }
 }
 
+mod block {
+    use std::io::Read;
+    use metatensor::{TensorBlock, TensorBlockRef};
+
+    const DATA_PATH: &str = "../../metatensor-core/tests/block.npz";
+
+    #[test]
+    fn load_file() {
+        let block = metatensor::io::load_block(DATA_PATH).unwrap();
+        check_block(block.as_ref());
+
+        let block = TensorBlock::load(DATA_PATH).unwrap();
+        check_block(block.as_ref());
+    }
+
+    #[test]
+    fn load_buffer() {
+        let mut file = std::fs::File::open(DATA_PATH).unwrap();
+        let mut buffer = Vec::new();
+        file.read_to_end(&mut buffer).unwrap();
+
+        let block = metatensor::io::load_block_buffer(&buffer).unwrap();
+        check_block(block.as_ref());
+
+        let block = TensorBlock::load_buffer(&buffer).unwrap();
+        check_block(block.as_ref());
+    }
+
+    #[test]
+    fn save_buffer() {
+        let mut file = std::fs::File::open(DATA_PATH).unwrap();
+        let mut buffer = Vec::new();
+        file.read_to_end(&mut buffer).unwrap();
+
+        let block = metatensor::io::load_block_buffer(&buffer).unwrap();
+
+        let mut saved = Vec::new();
+        metatensor::io::save_block_buffer(block.as_ref(), &mut saved).unwrap();
+        assert_eq!(buffer, saved);
+
+        saved.clear();
+        block.save_buffer(&mut saved).unwrap();
+        assert_eq!(buffer, saved);
+    }
+
+
+    fn check_block(block: TensorBlockRef) {
+        assert_eq!(block.values().as_array().shape(), [9, 5, 3]);
+        assert_eq!(block.samples().names(), ["system", "atom"]);
+        assert_eq!(block.components().len(), 1);
+        assert_eq!(block.components()[0].names(), ["o3_mu"]);
+        assert_eq!(block.properties().names(), ["n"]);
+
+        assert_eq!(block.gradient_list(), ["positions"]);
+        let gradient = block.gradient("positions").unwrap();
+        assert_eq!(gradient.values().as_array().shape(), [59, 3, 5, 3]);
+        assert_eq!(gradient.samples().names(), ["sample", "system", "atom"]);
+        assert_eq!(gradient.components().len(), 2);
+        assert_eq!(gradient.components()[0].names(), ["xyz"]);
+        assert_eq!(gradient.components()[1].names(), ["o3_mu"]);
+        assert_eq!(gradient.properties().names(), ["n"]);
+    }
+}
 
 mod labels {
     use std::io::Read;
     use metatensor::Labels;
 
+    const DATA_PATH: &str = "../../metatensor-core/tests/keys.npy";
+
     #[test]
     fn load_file() {
-        let labels = metatensor::io::load_labels("../../metatensor-core/tests/keys.npy").unwrap();
+        let labels = metatensor::io::load_labels(DATA_PATH).unwrap();
         check_labels(&labels);
 
-        let labels = Labels::load("../../metatensor-core/tests/keys.npy").unwrap();
+        let labels = Labels::load(DATA_PATH).unwrap();
         check_labels(&labels);
     }
 
     #[test]
     fn load_buffer() {
-        let mut file = std::fs::File::open("../../metatensor-core/tests/keys.npy").unwrap();
+        let mut file = std::fs::File::open(DATA_PATH).unwrap();
         let mut buffer = Vec::new();
         file.read_to_end(&mut buffer).unwrap();
 
@@ -94,7 +161,7 @@ mod labels {
 
     #[test]
     fn save_buffer() {
-        let mut file = std::fs::File::open("../../metatensor-core/tests/keys.npy").unwrap();
+        let mut file = std::fs::File::open(DATA_PATH).unwrap();
         let mut buffer = Vec::new();
         file.read_to_end(&mut buffer).unwrap();